+++ src/main/java/com/example/springbootstudy/database/repository/UserInfoRepository.java
@@ -3,6 +3,6 @@ package com.example.springbootstudy.database.repository;
 import com.example.springbootstudy.database.entity.UserInfo;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserInfoRepository extends CrudRepository<UserInfo, Integer> {
+public interface UserInfoRepository extends CrudRepository<UserInfo, Long> {
 
 }
