+++ src/main/java/com/example/springbootstudy/database/repository/UserLoginHistoryRepository.java
@@ -3,6 +3,6 @@ package com.example.springbootstudy.database.repository;
 import com.example.springbootstudy.database.entity.UserLoginHistory;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserLoginHistoryRepository extends CrudRepository<UserLoginHistory, Integer> {
+public interface UserLoginHistoryRepository extends CrudRepository<UserLoginHistory, Long> {
 
 }
